Version 1.5.6 of May 1, 2025 for irwin_v5 (fix efficiency for beta's)
The code for computing the beta(m+1)'s with the help of a @parallel
decorated procedure was done a bit hastily and had a big defect to
create child processes inheriting tasks of a (significantly) decreasing
duration, so that the global execution time ended up being the longest
one of the child processes which was quite worse than what slices
with maxworkers steps would have given.  This has been fixed.

The bug was somewhat hidden when showtimes=True was passed to the
irwin() or irwinpos() call because slices had length only about 1000, so
the difference between slowest and fastest parts proved small.  But the
effect was present there too, and with this commit, execution times are
improved also with showtimes=True.

Anyway, the computation of the beta(m+1)'s proves to take less or about
of the order of 10% of the total execution time, so this does not change
enormously the global execution times when requiring tens of thousands
of digits.