Fix rST mark-up for hyperlinks
The Docutils reference and documentation on hyperlinks are obscure, the
Sphinx reStructuredText primer is even worse on this and is even
misleading, and as a result almost noone in the world except the
Docutils maintainers have understood their mark-up for hyperlinks.

Besides, Docutils had a bug of not actually enforcing their own policies
and warned only in some cases that using a single final underscore after
the backtick creates completely unexpected

    Duplicate explicit target name

in case of identical link texts.  This bug (at 0.21.2) is resolved in
the Docutils 0.22 development branch, so it was needed (not really, but
let's be conscientious) to fix the mark-up to avoid such warnings in
future during the Sphinx build of the Page deployment.

And to add to all this, after all those years I don't even know if I
have to abridge reStructuredText as reST or rST.

Refers: https://github.com/sphinx-doc/sphinx/issues/13613
Fix rST mark-up  for hyperlinks